backend/requirements.txt
# Django 4.2 LTS supports Python 3.9+. Use Django 5.2+ when you upgrade to Python 3.10+.
Django>=4.2,<5
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.6
djangorestframework-camel-case>=1.4
openpyxl>=3.1
Pillow>=10.0
